requirements.txt
asgiref==3.5.2
backports.zoneinfo
Django==4.1.3
dnspython==2.2.1
flake8==5.0.4
mccabe==0.7.0
pycodestyle==2.9.1
pyflakes==2.5.0
pymongo==4.3.2
sqlparse==0.4.3

pandas~=1.3.5
Pillow~=9.2.0
tensorflow~=2.10.0
keras~=2.10.0
numpy~=1.21.5
opencv-python~=4.6.0.66
scipy~=1.8.1
sklearn~=0.0.post1
scikit-learn~=1.1.3